vendcare-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
stripe==8.5.0
pydantic==2.6.3
pydantic[email]>=2.6.3
email-validator==2.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
pymongo==4.6.3
motor==3.3.2
beanie==1.25.0
pydantic-settings>=2.0.0